+++ src/app/_components/MemoListModal.tsx
@@ -4,7 +4,7 @@ import Modal from '@/components/Modal'
 import Button from '@/components/Button'
 import { Icon } from '@/components/Icon'
 import { Fragment, useState } from 'react'
-import MemoListItem from '@/app/retrospects/create/[[...memoId]]/_components/MemoListItem'
+import MemoListItem from '@/app/retrospects/create/[[...id]]/_components/MemoListItem'
 import { useMemoListModalStore } from '@/store/memoListModal'
 import useMyMemoListQuery from '@/app/_queries/useMyMemoListQuery'
 import { cn } from '@/utils/cn'
