+++ src/app/retrospects/create/[[...id]]/_components/SubmitHeader/MemoSaveButton.tsx
@@ -2,13 +2,13 @@ import Button from '@/components/Button'
 import anyTrue from '@/utils/anyTrue'
 import { Icon } from '@/components/Icon'
 import { cn } from '@/utils/cn'
-import type { MutationResponseUnion } from '@/app/retrospects/create/[[...memoId]]/_types/retrospect'
+import type { MutationResponseUnion } from '../../_types/retrospect'
 import openCustomToast from '@/utils/openCustomToast'
 import handleSubmit from './handleSubmit'
 import useInvalidateQueries from '@/hooks/useInvalidateQueries'
 import { BasePayload } from './types'
-import useMemoCreateMutation from '@/app/retrospects/create/[[...memoId]]/_queries/useMemoCreateMutation'
-import useMemoUpdateMutation from '@/app/retrospects/create/[[...memoId]]/_queries/useMemoUpdateMutation'
+import useMemoCreateMutation from '../../_queries/useMemoCreateMutation'
+import useMemoUpdateMutation from '../../_queries/useMemoUpdateMutation'
 import { RefObject } from 'react'
 
 interface MemoSaveButtonProps {
